+++ packages/ui/cdk/src/lib/wallet-adapter-cdk.module.ts
@@ -1,8 +1,11 @@
 import { NgModule } from '@angular/core';
 import { HdWalletConnectButtonDirective } from './connect-button.directive';
+import { HdConnectWalletDirective } from './connect-wallet.directive';
 import { HdWalletDisconnectButtonDirective } from './disconnect-button.directive';
+import { HdDisconnectWalletDirective } from './disconnect-wallet.directive';
 import { HdWalletListItemComponent } from './list-item.component';
 import { HdObscureAddressPipe } from './obscure-address.pipe';
+import { HdSelectWalletDirective } from './select-wallet.directive';
 import { HdWalletAdapterDirective } from './wallet-adapter.directive';
 import { HdWalletIconComponent } from './wallet-icon.component';
 
@@ -14,6 +17,9 @@ import { HdWalletIconComponent } from './wallet-icon.component';
     HdWalletIconComponent,
     HdWalletAdapterDirective,
     HdWalletListItemComponent,
+    HdSelectWalletDirective,
+    HdConnectWalletDirective,
+    HdDisconnectWalletDirective,
   ],
   exports: [
     HdWalletConnectButtonDirective,
@@ -22,6 +28,9 @@ import { HdWalletIconComponent } from './wallet-icon.component';
     HdWalletIconComponent,
     HdWalletAdapterDirective,
     HdWalletListItemComponent,
+    HdSelectWalletDirective,
+    HdConnectWalletDirective,
+    HdDisconnectWalletDirective,
   ],
 })
 export class HdWalletAdapterCdkModule {}